update write method
